# Conflicts:
#	public/assets/js/controllers/main.js
#	public/views/main.html

+++ models/moodDay.js
@@ -2,7 +2,7 @@
 module.exports = function(sequelize, DataTypes) {
   var MoodDay = sequelize.define('MoodDay', {
     date: DataTypes.DATEONLY,
-    score: DataTypes.INTEGER,
+    score: DataTypes.FLOAT,
     updatedAt: DataTypes.DATE
   }, {
     classMethods: {
